+++ BackEnd/src/main/java/pe/edu/upc/backend/dtos/PromocionDTO.java
@@ -1,15 +1,15 @@
 package pe.edu.upc.backend.dtos;
 import java.time.LocalDate;
 public class PromocionDTO {
-    private int idPromocion;
+    private Long idPromocion;
     private String nombrePromocion;
     private String descripcionPromocion;
     private double precioPromocion;
     private double descuentoPromocion;
     private LocalDate fechaInicioPromocion;
     private LocalDate fechaFinPromocion;
-    public int getIdPromocion() { return idPromocion; }
-    public void setIdPromocion(int idPromocion) { this.idPromocion = idPromocion; }
+    public Long getIdPromocion() { return idPromocion; }
+    public void setIdPromocion(Long idPromocion) { this.idPromocion = idPromocion; }
     public String getNombrePromocion() { return nombrePromocion; }
     public void setNombrePromocion(String nombrePromocion) { this.nombrePromocion = nombrePromocion; }
     public String getDescripcionPromocion() { return descripcionPromocion; }
