+++ BackEnd/src/main/java/pe/edu/upc/backend/serviceimplements/PromocionImplement.java
@@ -17,9 +17,9 @@ public class PromocionImplement implements IPromocionService {
     @Override
     public void update(Promocion promocion) { pR.save(promocion); }
     @Override
-    public void delete(int idPromocion) { pR.deleteById(idPromocion); }
+    public void delete(Long idPromocion) { pR.deleteById(idPromocion); }
     @Override
-    public Promocion listById(int idPromocion) { return pR.findById(idPromocion).orElse(new Promocion()); }
+    public Promocion listById(Long idPromocion) { return pR.findById(idPromocion).orElse(new Promocion()); }
     @Override
     public List<Promocion> listarPorDescuentoMayor(double descuento) { return pR.listarPorDescuentoMayor(descuento); }
     @Override
