+++ BackEnd/src/main/java/pe/edu/upc/backend/serviceinterfaces/IComentarioService.java
@@ -6,8 +6,8 @@ public interface IComentarioService {
     public void insert(Comentario comentario);
     public List<Comentario> list();
     public void update(Comentario comentario);
-    public void delete(int idComentario);
-    public Comentario listId(int idComentario);
+    public void delete(Long idComentario);
+    public Comentario listId(Long idComentario);
     public List<Comentario> obtenerComentariosPorUsuario(int idUsuario);
     public List<Comentario> buscarComentariosPorNegocio(int idNegocio);
     public List<Comentario> buscarComentariosPorTurismo(int idTurismo);
